Move StructRecordCodegen to its own file
