+++ crates/burn-derive/src/record/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod codegen_struct;
 pub(crate) mod item;
 
 mod base;
